requirements.txt
aiogram>=3.22.0
openai>=1.40.0
python-dotenv>=1.0.1

# для локального RAG/ингеста (кэш "Обо мне" и FAQ)
numpy>=1.24.0
pypdf>=4.0.0
docx2txt>=0.8
tiktoken>=0.7.0

# webhook (если используете)
fastapi>=0.110.0
uvicorn>=0.29.0

# веб-поиск и загрузка страниц
ddgs>=1.7.7
httpx>=0.28.0
lxml>=5.3.0
